Added Portfolio class
